pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vidgen"
version = "0.1.0"
description = "Toolkit for fine-tuning video generative models such as AnimateDiff."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "VidGen Team" }
]
license = { text = "MIT" }
dependencies = [
  "accelerate>=0.26",
  "diffusers[torch]>=0.24.0",
  "transformers>=4.38",
  "torch>=2.1",
  "torchvision>=0.16",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "typer>=0.9",
  "rich>=13.7",
  "omegaconf>=2.3",
  "opencv-python>=4.9",
  "numpy>=1.26",
  "datasets>=2.16",
  "tqdm>=4.66"
]
keywords = ["video", "diffusion", "training", "fine-tuning"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1",
  "ruff>=0.3",
  "mypy>=1.8"
]

[project.scripts]
vidgen = "vidgen.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
minversion = "8.0"
pythonpath = ["src"]
addopts = "-ra"
